Add eslint and do what it says :D
Not using the the constructor shorthand is an interesting choice ...
probably for the best, but it did make me feel like I needed to add type
alises for Callback and FaliableCallback. That might be a good thing, I
can really talk about the continuation type as a representation of a
callback.